+++ e2e/draw.spec.ts
@@ -1,5 +1,6 @@
 import { test, expect } from '@playwright/test'
 import { openSnowbox } from './utils/openSnowbox'
+import { draw } from './utils/draw'
 
 const drawTargetId = '#vuex-target-draw-result'
 
@@ -9,15 +10,6 @@ test('the draw tool can be used to draw polygons on the map', async ({
 }) => {
   await openSnowbox(page)
 
-  const canvas = await page.locator('canvas')
-  const boundingBox = await canvas.boundingBox()
-  if (boundingBox === null) throw new Error('Canvas not found.')
-  const { width, height } = boundingBox
-  let { x, y } = boundingBox
-
-  x += width / 2
-  y += height / 2
-
   await page.getByLabel('Draw tools').click()
   await page.getByText('Draw, write and measure').click()
   await page.getByText('Polygon').click()
@@ -26,24 +18,16 @@ test('the draw tool can be used to draw polygons on the map', async ({
     // menu gets in the way in mobile
     await page.getByLabel('Draw tools').click()
     await expect(page.locator('.polar-draw-menu')).toHaveCount(0)
-    // canvas not fully visible initially in mobile
-    await canvas.scrollIntoViewIfNeeded()
-    const scrollY = await page.evaluate(() => window.scrollY)
-    y -= scrollY
   }
 
-  const moves: [number, number, string][] = [
+  await draw(page, [
     [0, 0, 'click'],
     [40, -40, 'click'],
     [40, 40, 'click'],
     [-80, 80, 'click'],
     [-80, -80, 'click'],
     [40, -40, 'dblclick'],
-  ]
-
-  for (const [xMove, yMove, method] of moves) {
-    await page.mouse[method]((x += xMove), (y += yMove))
-  }
+  ])
 
   const drawing = JSON.parse(await page.locator(drawTargetId).innerText())