+++ packages/clients/diplan/src/store/updateState/index.ts
@@ -15,6 +15,33 @@ import { enrichWithMetaServices } from './enrichWithMetaServices'
 let abortController: AbortController | null = null
 export const drawFeatureCollectionSource = 'plugin/draw/featureCollection'
 
+const reviseFeatureCollection = (
+  sourceFeatureCollection: FeatureCollection<GeometryType>
+) => {
+  let revisedFeatureCollection = cloneFeatureCollection(sourceFeatureCollection)
+
+  revisedFeatureCollection = {
+    ...revisedFeatureCollection,
+    features: revisedFeatureCollection.features.reduce(
+      (accumulator, feature) => {
+        if (['Polygon', 'MultiPolygon'].includes(feature.geometry.type)) {
+          accumulator.push(
+            ...unkinkPolygon(
+              cleanCoords(feature) as GeoJsonFeature<Polygon | MultiPolygon>
+            ).features
+          )
+        } else {
+          accumulator.push(cleanCoords(feature))
+        }
+        return accumulator
+      },
+      [] as GeoJsonFeature<GeometryType>[]
+    ),
+  }
+
+  return revisedFeatureCollection
+}
+
 // FeatureCollection is compatible to stupid clone
 export const cloneFeatureCollection = (
   // No GeometryCollection from Draw, hence the <GeometryType>
@@ -38,40 +65,25 @@ export const updateState = async ({
   const thisController = (abortController = new AbortController())
 
   // clone to prevent accidentally messing with the draw tool's data
-  let revisedFeatureCollection = cloneFeatureCollection(
-    rootGetters[drawFeatureCollectionSource]
-  )
-
-  revisedFeatureCollection = {
-    ...revisedFeatureCollection,
-    features: revisedFeatureCollection.features.reduce(
-      (accumulator, feature) => {
-        if (['Polygon', 'MultiPolygon'].includes(feature.geometry.type)) {
-          accumulator.push(
-            ...unkinkPolygon(feature as GeoJsonFeature<Polygon | MultiPolygon>)
-              .features
-          )
-        } else {
-          accumulator.push(feature)
-        }
-        return accumulator
-      },
-      [] as GeoJsonFeature<GeometryType>[]
-    ),
+  let revisedFeatureCollection
+  try {
+    revisedFeatureCollection = reviseFeatureCollection(
+      rootGetters[drawFeatureCollectionSource]
+    )
+  } catch {
+    // if this revision fails, errors are catastrophic and unfixable
+    commit('setSimpleGeometryValidity', false)
+    commit('setRevisionInProgress', false)
+    console.warn(
+      `@polar/client-diplan: Geometries entered were not valid. This may e.g. result from pulling points of a polygon in edit mode together until it's point-shaped.`
+    )
+    return
   }
 
   // merge first; relevant for both follow-up steps
   if (getters.configuration.mergeToMultiGeometries) {
     // TODO: turf provides "union" and "combine" methods, probably just use them
     revisedFeatureCollection = mergeToMultiGeometries(revisedFeatureCollection)
-  }
-
-  // removes duplicate points from polygons
-  revisedFeatureCollection = {
-    ...revisedFeatureCollection,
-    features: revisedFeatureCollection.features.map((feature) =>
-      cleanCoords(feature)
-    ),
   }
 
   if (getters.configuration.validateGeoJson) {