+++ packages/plugins/LayerChooser/CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## unpublished
+
+- Feature: Change configuration of `layer.type` to be set to `background` or any other String. All layers that have their `type` not set to `background` are considered "overlays". For the type `mask` locales are already implemented. For any additional type, these have to be added separately.
+- Fix: Layer options layer names are now translated if they're locale keys.
+
+## 2.1.0
+
+- Feature: Add new configuration parameter `component` to be able to interchange the UI component of this plugin.
+- Feature: Expose `Options.vue` as new export `LayerChooserOptions` in order to use it in custom implementations.
+- Feature: Expose `LayerWrapper.vue` as new export `LayerChooserLayerWrapper` in order to use it in custom implementations.
+- Fix: Resolve a bug where keyboard navigation in radio groups didn't work.
+- Chore: The documentation for `minZoom` and `maxZoom` has been updated regarding their inclusive interpretation of the zoom value.
+
 ## 2.0.0
 
 - Breaking: Upgrade `@masterportal/masterportalapi` from `2.40.0` to `2.45.0` and subsequently `ol` from `^9.2.4` to `^10.3.1`.